// types.ts
export type ViewState = 
  | 'auth'
  | 'home' 
  | 'analytics' 
  | 'teams' 
  | 'create-team'
  | 'profile' 
  | 'add-training' 
  | 'add-training-skiing'
  | 'activity-select' 
  | 'team-details' 
  | 'exercise-details' 
  | 'jump-details'
  | 'body-metrics'
  | 'all-body-metrics'
  | 'all-sessions'
  | 'coach-event-details'; // New View

export type UnitSystem = 'metric' | 'imperial';
// ...
export type JumpType = 'squat_jump' | 'cm_jump' | 'drop_jump' | '45s_jump' | 'single_leg_left' | 'single_leg_right';

export interface JumpLog {
  id: string;
  date: string;
  type: JumpType;
  value: number; // Stored in cm
}